backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Data Processing
polars==0.20.3
duckdb==0.9.2

# Async and Task Queue
celery==5.3.4
redis==5.0.1

# Security
cryptography==42.0.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Development
python-dotenv==1.0.0

# AI Providers
google-generativeai==0.3.2
google-cloud-aiplatform==1.38.1
openai==1.12.0
anthropic==0.18.0
# DeepSeek uses OpenAI-compatible API, no separate SDK needed

